1_ingestion/requirements.txt
kafka-python==2.0.2
numpy<2.0.0
pandas==2.1.1
pyarrow==13.0.0
s3fs